% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{buildPanel_oneCluster}
\alias{buildPanel_oneCluster}
\title{This UNTESTED function finds the best small marker panel for marking a single cluster, using
  proportion difference as the metric for determining the starting panel.}
\usage{
buildPanel_oneCluster(mapDat, clustersF, medianDat = NA, propIn = NA,
  clust = as.character(clustersF[1]), subSamp = NA, seed = 10,
  maxSize = 20, dexCutoff = 0.001, topGeneCount = 100)
}
\arguments{
\item{mapDat}{normalized data of the mapping (=reference) data set.}

\item{clustersF}{cluster calls for each cell.}

\item{medianDat}{median value for each leaf}

\item{propIn}{proportions of cells with expression > 1 in each leaf}

\item{clust}{which cluster to target?}

\item{subSamp}{number of random nuclei to select from each cluster, EXCEPT the target cluster;
set as NA to not subsample}

\item{seed}{for reproducibility}

\item{maxSize}{maximum size of marker gene panel}

\item{dexCutoff}{criteria for stopping: when improvement in fraction of cells properly mapped
dips below this value}

\item{topGeneCount}{number of top genes by proportion to consider}
}
\value{
a matrix of the top marker genes for each cluster.  Output matrix includes five columns:
  clust = cluster; panel = ordered genes in the panel for that cluster; onCorrect = fraction of
  correctly assigned cells in cluster; offCorrect = fraction of cells correctly assigned outside
  of cluster; dexTotal = additional dex explained by last gene added.
}
\description{
This UNTESTED function finds the best small marker panel for marking a single cluster, using
  proportion difference as the metric for determining the starting panel.
}
