% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{buildQualityTable}
\alias{buildQualityTable}
\title{Correct mapping at different tree heights}
\usage{
buildQualityTable(
  orderedGenes,
  dend,
  mapDat,
  medianDat,
  clustersF,
  minVal = 2,
  heights = c((0:100)/100),
  verbose = FALSE
)
}
\arguments{
\item{orderedGenes}{an ordered list of input genes (e.g. from an iterative algorithm)}

\item{dend}{dendrogram for mapping.}

\item{mapDat}{normalized data of the mapping (=reference) data set.}

\item{medianDat}{median value for each leaf}

\item{clustersF}{cluster calls for each cell}

\item{minVal}{minimum number of genes to consider from the list in the mapping}

\item{heights}{height in the tree to look at}

\item{verbose}{whether or not to show progress in the function}
}
\value{
a matrix of fractions of cells correctly mapped for different tree heights (columns)
  and different gene panels (rows)
}
\description{
This function takes as input an ordered set of marker genes (e.g., from at iterative algorithm,
  and returns an table showing the fraction of cells correctly mapped to a similar cell type
  (as defined by the heights parameter).  A height of 1 indicates correct mapping to the leaf.
}
