% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfishMapping.r
\name{buildTreeFromGenePanel}
\alias{buildTreeFromGenePanel}
\title{Build and plot dendrogram from gene panel}
\usage{
buildTreeFromGenePanel(dend = NA, refDat = NA, mapDat = refDat,
  medianDat = NA, requiredGenes = 2, clusters = NA,
  mappedAsReference = FALSE, genesToMap = rownames(mapDat),
  plotdendro = TRUE, returndendro = TRUE, mar = c(12, 5, 5, 5),
  main = NULL, ylab = NULL, use = "p", ...)
}
\arguments{
\item{dend}{dendrogram for mapping.  Ignored if medianDat is passed}

\item{refDat}{normalized data of the REFERENCE data set.  Ignored if medianDat is passed}

\item{mapDat}{normalized data of the MAPPING data set.  Default is to map the data onto itself.}

\item{medianDat}{representative value for each leaf and node.  If not entered, it is calculated}

\item{requiredGenes}{minimum number of genes required to be expressed in a cluster (column
of medianDat) for the cluster to be included (default=2)}

\item{clusters}{cluster calls for each cell}

\item{mappedAsReference}{if TRUE, returns the fraction of cells mapped to a node which are
were orginally clustered from that node; if FALSE (default) returns the fraction of cells
clustered under a node which are mapped to the correct node.}

\item{genesToMap}{which genes to include in the correlation mapping}

\item{plotdendro}{should the dendrogram be plotted (default = TRUE)}

\item{mar}{margins (for use with par)}

\item{main, ylab}{add title and labels to plot (default is NULL)}

\item{use, ...}{additional parameters for cor}

\item{returnDendro}{should the dendrogram be returned (default = TRUE)}
}
\value{
a list where the first entry is the resulting tree and the second entry is the
  fraction of cells correctly mapping to each node using the inputted gene panel.
}
\description{
Build and plot a dendrogram using correlation-based average linkage hierarchical
  clustering and only using a specified set of genes.  The output is the expected
  accuracy of mapping to each node in the tree, which gives an idea of the best-case
  expected results for mFISH analysis.
}
