% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfishMapping.r
\name{cellToClusterMapping_byCor}
\alias{cellToClusterMapping_byCor}
\title{Return top mapped correlation-based cluster and confidence}
\usage{
cellToClusterMapping_byCor(medianDat, mapDat, refDat = NA, clusters = NA,
  genesToMap = rownames(mapDat), use = "p", method = "p")
}
\arguments{
\item{medianDat}{representative value for each leaf and node.  If not entered, it is calculated}

\item{mapDat}{normalized data of the MAPPING data set.  Default is to map the data onto itself.}

\item{refDat}{normalized data of the REFERENCE data set.  Ignored if medianDat is passed}

\item{clusters}{cluster calls for each cell.  Ignored if medianDat is passed}

\item{genesToMap}{which genes to include in the correlation mapping}

\item{use}{additional parameter for cor (use="p" as default)}

\item{method}{additional parameter for cor (method="p" as default)}
}
\value{
data frame with the top match and associated correlation
}
\description{
Primary function for doing correlation-based mapping to cluster medians and also reporting the
  correlations and confidences.  This is wrapper for getTopMatch and corTreeMapping.
}
