% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{cellToClusterMapping_byRank}
\alias{cellToClusterMapping_byRank}
\title{Cell-based cluster mapping}
\usage{
cellToClusterMapping_byRank(
  mapDat,
  refDat,
  clustersF,
  genesToMap = rownames(mapDat),
  mergeFunction = rowMedians,
  useRank = TRUE,
  use = "p",
  method = "p"
)
}
\arguments{
\item{mapDat}{normalized data of the MAPPING data set.}

\item{refDat}{normalized data of the REFERENCE data set}

\item{clustersF}{factor indicating which cluster each cell type is actually assigned
to in the reference data set}

\item{genesToMap}{character vector of which genes to include in the correlation mapping}

\item{mergeFunction}{function for combining ranks; the tested choices are rowMeans or
rowMedians (default)}

\item{useRank}{use the rank of the correlation (default) or the correlation itself to
determine the top cluster}

\item{use}{additional parameter for cor (use='p' as default)}

\item{method}{additional parameter for cor (method='p' as default)}
}
\value{
a two column data matrix where the first column is the mapped cluster and the second
  column is a confidence call indicating how close to the top of the ranked list cells of the
  assigned cluster were located relative to their best possible location in the ranked list.
  This confidence score seems to be a bit more reliable than correlation at determining how
  likely a cell in a training set is to being correctly assigned to the training cluster.
}
\description{
Maps cells to clusters by correlating every mapped cell with every reference cell,
  ranking the cells by correlation, and the reporting the cluster with the lowest average rank.
}
