% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{corTreeMapping}
\alias{corTreeMapping}
\title{Primary function for doing correlation-based mapping to cluster medians.}
\usage{
corTreeMapping(dend = NA, refDat = NA, mapDat = refDat, medianDat = NA,
  clusters = NA, genesToMap = rownames(mapDat), use = "p", ...)
}
\arguments{
\item{dend}{dendrogram for mapping.  Ignored if medianDat is passed}

\item{refDat}{normalized data of the REFERENCE data set.  Ignored if medianDat is passed}

\item{mapDat}{normalized data of the MAPPING data set.  Default is to map the data onto itself.}

\item{medianDat}{representative value for each leaf and node.  If not entered, it is calculated}

\item{clusters}{cluster calls for each cell.  Ignored if medianDat is passed}

\item{genesToMap}{which genes to include in the correlation mapping}

\item{use, ...}{additional parameters for cor}
}
\value{
matrix with the correlation between expression of each cell and representative value for each node and leaf
}
\description{
Primary function for doing correlation-based mapping to cluster medians.
}
