% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{distTreeMapping}
\alias{distTreeMapping}
\title{Returns the distance between expression of each cell and representative value for each node and
  leaf (default is based on euclidean distance).  In our hands this is does not work very well.}
\usage{
distTreeMapping(dend = NA, refDat = NA, mapDat = refDat, medianDat = NA,
  clusters = NA, genesToMap = rownames(mapDat), returnSimilarity = TRUE,
  use = "p", ...)
}
\arguments{
\item{dend}{dendrogram for mapping.  Ignored if medianDat is passed}

\item{refDat}{normalized data of the REFERENCE data set.  Ignored if medianDat is passed}

\item{mapDat}{normalized data of the MAPPING data set.  Default is to map the data onto itself.}

\item{medianDat}{representative value for each leaf and node.  If not entered, it is calculated}

\item{clusters}{cluster calls for each cell.  Ignored if medianDat is passed}

\item{genesToMap}{which genes to include in the correlation mapping}

\item{returnSimilarity}{FALSE to return distance, TRUE to return something like a similarity}

\item{use, ...}{additional parameters for dist (for back-compatiblity; doesn't work)}
}
\description{
Returns the distance between expression of each cell and representative value for each node and
  leaf (default is based on euclidean distance).  In our hands this is does not work very well.
}
