% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{filterByClass}
\alias{filterByClass}
\title{Filter by meta-data}
\usage{
filterByClass(
  classVector,
  sampleInfo,
  classColumn = "cluster_type_label",
  clusterColumn = "cluster_label",
  threshold = 0.1
)
}
\arguments{
\item{classVector}{vector corresponding to the class information for filtering (e.g., vector
of label calls)}

\item{sampleInfo}{matrix of sample information with rows corresponding to cells and columns
corresponding to meta-data}

\item{classColumn}{column name of class information}

\item{clusterColumn}{column name of cluster information}

\item{threshold}{minimum fraction of cluster cells from a given class to be considered present}
}
\value{
a matrix of filters with rows as clusters and columns as classes with entries of TRUE or
  FALSE indicating whether cells from a given class can assigned to that cluster, given threshold.
}
\description{
Return a filter of TRUE/FALSE values for a given piece of meta-data (e.g., broad class).
}
