% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfishMapping.r
\name{filterCells}
\alias{filterCells}
\title{Filter cells}
\usage{
filterCells(datIn, kpSamp)
}
\arguments{
\item{datIn}{a fishScaleAndMap output list}

\item{kpSamp}{a vector of samples to keep (of the same length as the number of cells in datIn)}
}
\value{
filtered fishScaleAndMap object
}
\description{
Remove a select set of cells from all elements of a fishScaleAndMap output
}
