% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{filterPanelGenes}
\alias{filterPanelGenes}
\title{Filter genes for spatial transcriptomics panel}
\usage{
filterPanelGenes(summaryExpr, propExpr = summaryExpr,
  onClusters = 1:dim(summaryExpr)[2], offClusters = NULL,
  geneLengths = NULL, startingGenes = c("GAD1", "SLC17A7"),
  numBinaryGenes = 500, minOn = 10, maxOn = 250, maxOff = 50,
  minLength = 960, fractionOnClusters = 0.5, excludeGenes = NULL,
  excludeFamilies = c("LOC", "LINC", "FAM", "ORF", "KIAA", "FLJ", "DKFZ",
  "RIK", "RPS", "RPL", "\\\\-"))
}
\arguments{
\item{summaryExpr}{Matrix of summarized expression levels for a given cluster.  Typically the median
or mean should be used.  Rows are genes and columns are samples.  ROW NAMES MUST BE GENE SYMBOLS!}

\item{propExpr}{Proportion of cells expressed in each cluster for use with binary score calculation
(default = summaryExpr, which is not recommended)}

\item{onClusters}{Vector indicating which clusters should be included in the gene panel (default
is all clusters. Can be logical or numeric, or a character string of cluster names)}

\item{offClusters}{Vector indidicating from which clusters expression should be avoided}

\item{numBinaryGenes}{Number of genes to include in the final panel.  Genes are sorted by binary
score using `getBetaScore` and this number of genes are chosen (default = 500)}

\item{minOn}{Minimum summary expression level in most highly expressed "on" cluster (default = 10)}

\item{maxOn}{Maximum summary expression level in most highly expressed "on" cluster (default = 250)}

\item{maxOff}{Maximum summary expression level in most highly expressed "off" cluster (default = 50)}

\item{minLength}{Minimum gene length for marker gene selection.  Ignored if geneLength is not 
provided (default = 960)}

\item{fractionOnClusters}{What is the maximum fraction of clusters in which a gene can be expressed
(as defined by propExpr>0.5; default = 0.5).  This prevents nearly ubiquitous genes from selection}

\item{excludeGenes}{Which genes should be excluded from the analysis (default is none)}

\item{excludeFamilies}{Which gene classes or families should be excluded from the analysis?  More
specifically, any gene that contain these strings of characters anywhere in the symbol will be
excluded (default is "LOC","LINC","FAM","ORF","KIAA","FLJ","DKFZ","RIK","RPS","RPL","\\-").}

\item{geneLength}{Optional vector of gene lengths in same order as summaryExpr.  Default is NULL}
}
\value{
A character vector of genes meeting all constraints
}
\description{
Returns a set of genes for inclusion in a spatial transcriptomics panel based on a series of 
  hard-coded and user-defined constraints
}
