% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{fractionCorrectWithGenes}
\alias{fractionCorrectWithGenes}
\title{Fraction of cells correctly assigned}
\usage{
fractionCorrectWithGenes(
  orderedGenes,
  mapDat,
  medianDat,
  clustersF,
  verbose = FALSE,
  plot = TRUE,
  return = TRUE,
  ...
)
}
\arguments{
\item{orderedGenes}{an ordered list of input genes (e.g. from an iterative algorithm)}

\item{mapDat}{normalized data of the mapping (=reference) data set.}

\item{medianDat}{median value for each leaf}

\item{clustersF}{cluster calls for each cell}

\item{verbose}{whether or not to show progress in the function}

\item{plot}{if TRUE, plotCorrectWithGenes is run}

\item{return}{if TRUE, the value is returned}

\item{...}{parameters passed to plotCorrectWithGenes (if plot=TRUE)}
}
\value{
a vector showing the fraction of cells correctly mapped to each cluster
}
\description{
This function takes as input an ordered set of marker genes (e.g., from at iterative
  algorithm), and returns a vector showing the fraction of cells correctly mapped.
}
