% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{FscoreWithGenes}
\alias{FscoreWithGenes}
\title{Returns F-Score, Precision and Recall for cell classification}
\usage{
FscoreWithGenes(orderedGenes, mapDat, medianDat, realCluster, focusGroup)
}
\arguments{
\item{orderedGenes}{an ordered list of input genes (e.g. from an iterative algorithm)
#' @param mapDat normalized data of the mapping (=reference) data set.}

\item{medianDat}{median value for each leaf}

\item{realCluster}{real cluster assignements}

\item{focusGroup}{which group of clusters to calculate F-score, precision, recall for}

\item{foundCluster}{cluster assignments from classification method}
}
\value{
a list with the first entry showing the F-score as the function of the number of genes included and
 the second and third entry showing precision and recall
}
\description{
This function takes as input an ordered set of marker genes (e.g., from at iterative
  algorithm), and returns a vector showing the fraction of cells correctly mapped.
}
