% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{generateMultipleCellReferenceSet}
\alias{generateMultipleCellReferenceSet}
\title{Generate reference set of pseudo-cells}
\usage{
generateMultipleCellReferenceSet(refDat, clustersF,
  genesToUse = rownames(refDat), cellsPerMerge = 5, numberOfMerges = 10,
  mergeFunction = rowMedians, seed = 1)
}
\arguments{
\item{refDat}{normalized data of the REFERENCE data set}

\item{clustersF}{factor indicating which cluster each cell type is actually assigned to in the reference data set}

\item{cellsPerMerge}{Number of cells to include in each combo cell}

\item{numberOfMerges}{Number of combo cells to include per cell type}

\item{mergeFunction}{function for combining cells into combo cells (use rowMeans or rowMedians)}

\item{seed}{for resproducibility}

\item{genesToMap}{which genes to include in the correlation mapping}
}
\value{
list where first element is data matrix of multi-cells by genes and
  second element is a vector of corresponding clusters
}
\description{
Creates a new reference set as input for cellToClusterMapping_byRank, where each 'cell' is the
  combiniation of several cells and this is run several times using different subsets of cells.
}
