% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{get_subtree_label}
\alias{get_subtree_label}
\title{Gets subtree labels for lca function.}
\usage{
get_subtree_label(dend)
}
\arguments{
\item{dend}{a cluster dendrogram}
}
\value{
vector of subtree labels
}
\description{
Gets subtree labels for lca function.
}
