% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{getBetaScore}
\alias{getBetaScore}
\title{Returns a beta score which indicates the binaryness of a gene across clusters.  High scores
  (near 1) indicate that a gene is either on or off in nearly all cells of every cluster.
  Scores near 0 indicate a cells is non-binary (e.g., not expressed, ubiquitous, or
  randomly expressed).  This value is used for gene filtering prior to defining clustering.}
\usage{
getBetaScore(propExpr, returnScore = TRUE, spec.exp = 2)
}
\arguments{
\item{propExpr}{a matrix of proportions of cells (rows) in a given cluster (columns) with
CPM/FPKM > 1 (or 0, HCT uses 1)}

\item{returnScore}{if TRUE returns the score, if FALSE returns the ranks}

\item{spec.exp}{scaling factor (recommended to leave as default)}
}
\value{
returns a numeric vector of beta score (or ranks)
}
\description{
Returns a beta score which indicates the binaryness of a gene across clusters.  High scores
  (near 1) indicate that a gene is either on or off in nearly all cells of every cluster.
  Scores near 0 indicate a cells is non-binary (e.g., not expressed, ubiquitous, or
  randomly expressed).  This value is used for gene filtering prior to defining clustering.
}
