% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{getBranchList}
\alias{getBranchList}
\title{Branch list}
\usage{
getBranchList(dend, branches = list(), allTips = as.character(dend \%>\%
  labels))
}
\arguments{
\item{dend}{dendrogram for mapping.  Ignored if medianDat is passed}

\item{branches}{do not change from default}

\item{allTips}{do not change from default}
}
\value{
a list of branch information for use with leafToNodeMedians
}
\description{
Returns branches of a dendrogram in a specific format
}
