% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{getConfusionMatrix}
\alias{getConfusionMatrix}
\title{Confusion matrix}
\usage{
getConfusionMatrix(realCluster, foundCluster, proportions = TRUE)
}
\arguments{
\item{realCluster}{character vector of assigned clusters}

\item{foundCluster}{character vector of mapped clusters}

\item{proportions}{FALSE if the counts are to be returned and TRUE if the proportions are to be returned}
}
\description{
Returns a confusion matrix of the found (mapped) vs. real (assigned) clusters.
}
