% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfishMapping.r
\name{getDend}
\alias{getDend}
\title{Build a dendrogram from gene panel}
\usage{
getDend(dat, distFun = function(x) return(as.dist(1 - WGCNA::cor(x))), ...)
}
\arguments{
\item{dat}{matrix of values (e.g., genes x clusters) for calculating the dendrogram}

\item{distFun}{function for calculating distance matrix (default is correlation-based)}

\item{...}{additional variables for distFun}
}
\value{
dendrogram
}
\description{
Build a dendrogram from an inputted data matrix.
}
