% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{getTopMatch}
\alias{getTopMatch}
\title{Get top leaf match}
\usage{
getTopMatch(memb.cl)
}
\arguments{
\item{memb.cl}{membership scores for each leaf}
}
\value{
a matrix where first column is found cluster and second column is confidence score
}
\description{
Returns the top leaf match for each cell and the corresponding fraction mapping there.
}
