% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfishMapping.r
\name{labelDend}
\alias{labelDend}
\title{Label dendrogram nodes}
\usage{
labelDend(dend, n = 1)
}
\arguments{
\item{dend}{dendrogram object}

\item{distFun}{starting numeric node value (default=1)}
}
\value{
a list where the first item is the new dendrogram object and the second item is the final
  numeric node value
}
\description{
Add numeric node labels to a dendrogram.
}
