% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{layerScale}
\alias{layerScale}
\title{Fraction of cells per layer}
\usage{
layerScale(layerIn, layerNm = c("L1", "L2/3", "L4", "L5", "L6"), scale = TRUE)
}
\arguments{
\item{layerIn}{a list corresponding to all layers of dissection for a given sample}

\item{layerNm}{names of all layers.  set to NULL to have this calculated}

\item{scale}{if TRUE (default), scale to the total number of cells}
}
\value{
vector indicating the fraction of cells in each layerNm layer
}
\description{
Determines the expected proportions in each layer based on input
}
