% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{lca}
\alias{lca}
\title{Get lowest common ancestor (defined cluster pairs)}
\usage{
lca(dend, l1, l2, l = rep(attr(dend, "label"), length(l1)))
}
\arguments{
\item{dend}{a cluster dendrogram}

\item{l1}{a vector of node labels}

\item{l2}{a second fector of node labels (of the same length as l1)}

\item{l}{do not adjust; required for recursive function}
}
\value{
The function will return a vector for lowest common ancestor for every pair of nodes in l1 and l2
}
\description{
Maps a cluster back up the tree to the first node where the mapped and correct clusters agree.
}
