% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{makeLCAtable}
\alias{makeLCAtable}
\title{Calculates the vector for lowest common ancestor for every pair of leaves in a tree and returns a
  vector in a specific format for faster look-up.}
\usage{
makeLCAtable(dend, includeInternalNodes = FALSE, verbose = FALSE)
}
\arguments{
\item{dend}{a cluster dendrogram}

\item{includeInternalNodes}{should internal nodes be included in the output?}

\item{verbose}{if TRUE, status will be printed to the screen, since function is relatively slow
for large trees (default FALSE)}
}
\value{
The function will return a vector for lowest common ancestor for every pair of leaves
  in dend.  Vector names are l1|||l2 for string parsing in other functions.
}
\description{
Calculates the vector for lowest common ancestor for every pair of leaves in a tree and returns a
  vector in a specific format for faster look-up.
}
