% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfishMapping.r
\name{mergeFish}
\alias{mergeFish}
\title{Merge two fishScaleAndMap objects}
\usage{
mergeFish(datFish1, datFish2)
}
\arguments{
\item{datFish1}{a fishScaleAndMap output list}

\item{datFish2}{a second fishScaleAndMap output list.}
}
\value{
a new fishScaleAndMap output list with the two original ones merged
}
\description{
Merges all components of two fishScaleAndMap objects to create a new 
  one. Note: only meta-data and mappingResults that is present in BOTH 
  objects will be returned.
}
