% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\docType{package}
\name{mfishtools}
\alias{mfishtools}
\title{mfishtools: Building Gene Sets and Mapping mFISH Data.}
\description{
This repository includes code for gene selection for spatial transcriptomics methods and for
mapping of spatial transcriptomics (or RNA-Seq data) onto a RNA-Seq reference. Specific topics include:
1) Correlation-based mapping of cells to reference cell types
2) Iterative building of gene panels a greedy algorithm with pre-defined constraints
3) Visualizations related to gene mapping a gene panel selection
}
