% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{outputTopConfused}
\alias{outputTopConfused}
\title{Table of confused clusters}
\usage{
outputTopConfused(confusionProp, count = 10)
}
\arguments{
\item{confusionProp}{confusion matrix (e.g., output from getConfusionMatrix).}

\item{count}{number of top confusions to show}
}
\value{
a 3 x count matrix of the top confused pairs of clusters with the three columns corresponding
  to mapped cluster, assigned cluster, and fraction of cells incorrectly mapped, respectively.
}
\description{
This function returns a table of the top confused clusters (assigned clusters incorrectly mapped)
}
