% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{plotConfusionVsConfidence}
\alias{plotConfusionVsConfidence}
\title{Confusion plot vs. confidence}
\usage{
plotConfusionVsConfidence(
  foundClusterAndScore,
  realCluster,
  RI = (31:100)/100,
  main = "\% mapping (blue) / correct (orange)",
  ylab = "Percent",
  xlab = "Fraction correctly mapped to leaf",
  type = "l",
  xlim = range(RI),
  ...
)
}
\arguments{
\item{foundClusterAndScore}{matrix where first column is found cluster and second column is
confidence score (e.g., output from getTopMatch)}

\item{realCluster}{character vector of assigned clusters}

\item{...}{additional parameters for the plot function}
}
\description{
Produces line plots showing the percent of correctly mapped cells above a certain confidence value (or score).
  This is a wrapper for plot.
}
