% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{plotCorrectWithGenes}
\alias{plotCorrectWithGenes}
\title{Plot fraction correct}
\usage{
plotCorrectWithGenes(
  frac,
  genes = names(frac),
  xlab = "Number of genes in panel",
  main = "All clusters gene panel",
  ylim = c(-10, 100),
  lwd = 5,
  ylab = "Percent of nuclei correctly mapping",
  colLine = "grey",
  ...
)
}
\arguments{
\item{frac}{a numeric vector indicating the fraction of cells correctly mapped for a given gene panel}

\item{genes}{ordered character vector (e.g., of genes) to be plotted; default is names(frac)}

\item{...}{additional parameters for plot.}
}
\description{
This function is a wrapper for plot designd for plotting the fraction correctly mapped for a given gene set.
  If geneN is the Nth gene, the plotted value indicates correct mapping using genes 1:N.
}
