% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfishMapping.r
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot heatmap}
\usage{
plotHeatmap(datIn, group, groups = NULL, useScaled = FALSE,
  colors = rep("black", dim(datIn$mapDat)[2]), colormap = gray.colors,
  pch = 19, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{datIn}{a fishScaleAndMap output list}

\item{group}{a character vector (or factor) indicating how to split the data (e.g., cluster 
call) or a metadata/mappingResults column name}

\item{groups}{a character vector of groups to show (default is levels of group)}

\item{useScaled}{plot the scaled (TRUE) or unscaled (FALSE; default) values}

\item{colors}{a character vector (or factor) indicating how to color the plots (e.g., layer 
or gene expression) or a metadata/mappingResults column name (default is all black)}

\item{colormap}{function to use for the colormap for the data (default gray.colors)}

\item{pch, ...}{other parameters for plot}

\item{xlim, ylim}{for plot, but will be calculated if not entered}
}
\value{
filtered fishScaleAndMap object
}
\description{
Plot the heatmap of cells ordering by a specified order.  This is a wrapper for heatmap.2
}
