% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{plotNodes}
\alias{plotNodes}
\title{Plot dendrogram}
\usage{
plotNodes(
  tree,
  value = rep(1, length(labels(tree))),
  cexScale = 2,
  margins = c(10, 5, 2, 2),
  cols = "black",
  pch = 19,
  ...
)
}
\arguments{
\item{tree}{a dendrogram object}

\item{value}{numeric vector corresponding to the size of each node}

\item{cexScale}{a global cex multiplier for node sizes}

\item{margins}{set the margins using par(mar=margins)}

\item{cols}{vector of node colors (or a single value)}

\item{pch}{vector of node pch shapes (or a single value)}

\item{...}{additional parameters for the plot function}
}
\description{
Plots a dendrogram with set not colors, shapes, sizes and labels.  This is a wrapper for plot.
}
