% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfishMapping.r
\name{plotTsne}
\alias{plotTsne}
\title{Plot TSNE}
\usage{
plotTsne(
  datIn,
  colorGroup = "none",
  labelGroup = "none",
  useScaled = FALSE,
  capValue = Inf,
  perplexity = 10,
  theta = 0.5,
  main = "TSNE plot",
  maxNchar = 1000,
  seed = 10
)
}
\arguments{
\item{datIn}{a fishScaleAndMap output list}

\item{colorGroup}{a character vector (or factor) indicating how to color the Tsne (e.g., cluster
call) or a metadata/mappingResults column name (default=NULL)}

\item{labelGroup}{a character vector (or factor) indicating how to label the Tsne (e.g., cluster
call) or a metadata/mappingResults column name (default=NULL)}

\item{useScaled}{plot the scaled (TRUE) or unscaled (FALSE; default) values}

\item{capValue}{values above capValue will be capped at capValue (default is none)}

\item{perplexity, theta}{other parameters for Rtsne}

\item{main}{title of the plot}

\item{maxNchar}{what is the maximum number of characters to display in the plot for each entry?}

\item{seed}{for reproducibility}
}
\value{
Only returns if there is an error
}
\description{
Plot a TSNE of the data, with assigned colors and labels from provided variables.  Note that this
  function is a modification of code from Pabloc (https://www.r-bloggers.com/author/pabloc/) from
  https://www.r-bloggers.com/playing-with-dimensions-from-clustering-pca-t-sne-to-carl-sagan/
}
