% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{possibleClustersByPriors}
\alias{possibleClustersByPriors}
\title{This function will return a vector of possible clusters for cells that meet a set of priors for each layer}
\usage{
possibleClustersByPriors(cluster, layer, subsetVector = rep(TRUE,
  length/(cluster)), useClusters = sort(unique(cluster)), rareLimit = 0.005,
  layerNm = c("L1", "L2/3", "L4", "L5", "L6"), scaleByLayer = TRUE,
  smartWeight = TRUE, spillFactor = 0.15, weightCutoff = 0.02)
}
\arguments{
\item{cluster}{vector of all clusters}

\item{layer}{list of layers for each cluster entry (for data sets with only laminar dissections,
each list entry will be of length 1)}

\item{subsetVector}{a vector of TRUE/FALSE values indicated whether the entry is in the subset of
interest (e.g., Cre lines); default is all}

\item{useClusters}{a set of clusters to be considered a priori (e.g., GABA vs. glut); default is all}

\item{rareLimit}{define any values less than this as 0.  The idea is to exclude rare cells}

\item{layerNm}{names of all layers.  set to NULL to have this calculated}

\item{scaleByLayer}{if TRUE, scales to the proportion of cells in each layer}

\item{smartWeight}{if TRUE, multilayer dissections are weighted smartly by cluster, rather than evenly by cluster (FALSE)}

\item{spillFactor}{fractional amount of cells in a layer below which it is assumed no cells are from that layer in multilayer dissection}

\item{weightCutoff}{anything less than this is set to 0 for convenience}
}
\value{
a vector of possible clusters for cells that meet a set of priors for each layer
}
\description{
This function will return a vector of possible clusters for cells that meet a set of priors for each layer
}
