% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfishMapping.r
\name{quantileTruncate}
\alias{quantileTruncate}
\title{Quantile normalize, truncate, and scale}
\usage{
quantileTruncate(x, qprob = 0.9, maxVal = 1, truncate = TRUE, ...)
}
\arguments{
\item{x}{input data vector}

\item{qprob}{probs value to result from quantile (default=0.9)}

\item{maxVal}{max value for scaling (default=1)}

\item{truncate}{should data above the qprob threshold be truncated (default=yes)}

\item{...}{not used}
}
\value{
scaled vector
}
\description{
Quantile normalize, truncate, and scale a numeric vector (e.g. mFISH data from one gene)
}
