% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{resolve_cl}
\alias{resolve_cl}
\title{Tree-based mapping (internal)}
\usage{
resolve_cl(cl.g, cl.med, markers, dat, map.dat, select.cells, p = 0.7,
  low.th = 0.2)
}
\arguments{
\item{cl.g}{all clusters}

\item{cl.med}{cluster medians}

\item{markers}{gene markers}

\item{dat}{normalized data of the REFERENCE data set}

\item{map.dat}{normalized data of the MAPPING data set.  Default is to map the data onto itself.}

\item{select.cells}{which cells to use?}

\item{p}{proportion of marker genes to include in each iteration of the mapping algorithm.}

\item{low.th}{the minimum difference in Pearson correlation required to decide on which branch
to map to. otherwise, a random branch is chosen.}
}
\value{
a vector of the mapped cluster
}
\description{
Returns the mapped cluster call of each cell to each leaf. This function is called by map_dend
}
