% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{rowMedians}
\alias{rowMedians}
\title{Calculate row medians}
\usage{
rowMedians(x, ...)
}
\arguments{
\item{x}{a matrix or vector of values}
}
\value{
if the input is a matrix, return row medians.  If a vector, it returns the inputted vector
}
\description{
This is a wrapper for matrixStats::rowMedians that doesn't crash if a vector is provided as input
}
