% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{smartLayerAllocation}
\alias{smartLayerAllocation}
\title{Returns a numeric vector saying how to weight a particular cell for each layer, using a smart
  weighting strategy}
\usage{
smartLayerAllocation(layerIn, useLayer = "L1", spillFactor = 0.15,
  weightCutoff = 0.02, layerNm = c("L1", "L2/3", "L4", "L5", "L6"))
}
\arguments{
\item{layerIn}{a list corresponding to all layers of dissection for a given sample}

\item{useLayer}{target layer}

\item{spillFactor}{fractional amount of cells in a layer below which it is assumed no cells are
from that layer in multilayer dissection}

\item{weightCutoff}{anything less than this is set to 0 for convenience and to avoid rare types}

\item{layerNm}{names of all layers.  set to NULL to have this calculated}
}
\value{
numeric vector saying how to weight a particular cell for each layer, using a smart
  weighting strategy
}
\description{
Returns a numeric vector saying how to weight a particular cell for each layer, using a smart
  weighting strategy
}
