% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{subsampleCells}
\alias{subsampleCells}
\title{Subsample cells}
\usage{
subsampleCells(clusters, subSamp = 25, seed = 5)
}
\arguments{
\item{clusters}{vector of cluster labels (or any category) in factor or character format}

\item{subSamp}{maximum number of values for each category to subsample.  Can be single integer
for global subsampling, or a *named* vector corresponding to how many values to take from each
category in clusters.}

\item{seed}{for reproducibility}
}
\value{
returns a vector of TRUE / FALSE with a maximum of subSamp TRUE calls per category
}
\description{
Subsets a categorical vector to include up to a maximum number of values for each category.
}
