% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfishMapping.r
\name{subsetFish}
\alias{subsetFish}
\title{Subset fishScaleAndMap object}
\usage{
subsetFish(datFish, subset)
}
\arguments{
\item{datFish}{a fishScaleAndMap output list}

\item{subset}{a boolean or numeric vector of the elements to retain}
}
\value{
a fishScaleAndMap output list with updated scaledX and scaleY coordinates
}
\description{
Subsets all components in a fishScaleAndMap object
}
