% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfishMapping.r
\name{buildTreeFromGenePanel}
\alias{buildTreeFromGenePanel}
\title{Build and plot dendrogram from gene panel}
\usage{
buildTreeFromGenePanel(dend = NA, refDat = NA, mapDat = refDat,
  medianDat = NA, clusters = NA, genesToMap = rownames(mapDat),
  plotdendro = TRUE, mar = c(12, 5, 5, 5), use = "p", ...)
}
\arguments{
\item{dend}{dendrogram for mapping.  Ignored if medianDat is passed}

\item{refDat}{normalized data of the REFERENCE data set.  Ignored if medianDat is passed}

\item{mapDat}{normalized data of the MAPPING data set.  Default is to map the data onto itself.}

\item{medianDat}{representative value for each leaf and node.  If not entered, it is calculated}

\item{clusters}{cluster calls for each cell}

\item{genesToMap}{which genes to include in the correlation mapping}

\item{plotdendro}{should the dendrogram be plotted (default = TRUE)}

\item{mar}{margins (for use with par)}

\item{use, ...}{additional parameters for cor}
}
\value{
matrix with the correlation between expression of each cell and representative value for each node and leaf
}
\description{
Build and plot a dendrogram using correlation-based average linkage hierarchical clustering
  and only using a specified set of genes.  The output is the expected accuracy of mapping 
  to each node in the tree, which gives an idea of the best-case expected results for 
  mFISH analysis.
}
