% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{corTreeMapping}
\alias{corTreeMapping}
\title{Correlation-based cluster mapping}
\usage{
corTreeMapping(
  mapDat,
  medianDat,
  dend = NULL,
  refDat = NA,
  clusters = NA,
  genesToMap = rownames(mapDat),
  use = "p",
  method = "p"
)
}
\arguments{
\item{mapDat}{normalized data of the MAPPING data set.  Default is to map the data onto itself.}

\item{medianDat}{representative value for each leaf and node.  If not entered, it is calculated}

\item{dend}{dendrogram for mapping.  If provided, correlations to nodes are also returned}

\item{refDat}{normalized data of the REFERENCE data set.  Ignored if medianDat is passed}

\item{clusters}{cluster calls for each cell.  Ignored if medianDat is passed}

\item{genesToMap}{which genes to include in the correlation mapping}

\item{use}{additional parameter for cor (use='p' as default)}

\item{method}{additional parameter for cor (method='p' as default)}
}
\value{
matrix with the correlation between expression of each cell and representative value for
  each leaf and node
}
\description{
Primary function for doing correlation-based mapping to cluster medians.  This is wrapper for cor
  and returns a correlation matrix.
}
