% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfishMapping.r
\name{filterCells}
\alias{filterCells}
\title{Filter (subset) fishScaleAndMap object}
\usage{
filterCells(datFish, subset)
}
\arguments{
\item{datFish}{a fishScaleAndMap output list}

\item{subset}{a boolean or numeric vector of the elements to retain}
}
\value{
a fishScaleAndMap output subsetted to the requested elements
}
\description{
Subsets all components in a fishScaleAndMap object
}
