% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGenesAndMapping.r
\name{plotCorrectByType}
\alias{plotCorrectByType}
\title{Plot fraction correct for each cell type}
\usage{
plotCorrectByType(res, types = colnames(res), xlab = "Cell type",
  main = "Mapping performance for each cell type",
  ylab = "Percent of cells correctly mapping", ...)
}
\arguments{
\item{types}{cell types, default is names(frac)}

\item{...}{additional parameters for plot.}

\item{frac}{a numeric vector indicating the fraction of cells correctly mapped for each cell type}
}
\description{
This function is a wrapper for plot designed for plotting the fraction correctly mapped for each
cell type
}
