% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfishMapping.r
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot heatmap}
\usage{
plotHeatmap(datIn, group, groups = NULL, grouplab = "Grouping",
  useScaled = FALSE, capValue = Inf, colormap = grey.colors(1000),
  pch = 19, xlim = NULL, ylim = NULL, Rowv = FALSE, Colv = FALSE,
  dendrogram = "none", trace = "none", margins = c(6, 10),
  rowsep = NULL, sepwidth = c(0.4, 0.4), key = FALSE, ...)
}
\arguments{
\item{datIn}{a fishScaleAndMap output list}

\item{group}{a character vector (or factor) indicating how to order the heatmap (e.g., cluster
call) or a metadata/mappingResults column name}

\item{groups}{a character vector of groups to show (default is levels of group)}

\item{grouplab}{label for the grouping in the heatmap (default is 'Grouping' or the value for group)}

\item{useScaled}{plot the scaled (TRUE) or unscaled (FALSE; default) values}

\item{capValue}{values above capValue will be capped at capValue (default is none)}

\item{colormap}{set of values to use for the colormap for the data (default heat_colors)}

\item{Rowv, Colv, dendrogram, trace, margins, rowsep, colsep, key, ...}{other parameters for heatmap.2
(some default values are different)}
}
\value{
Only returns if there is an error
}
\description{
Plot the heatmap of cells ordering by a specified order.  This is a wrapper for heatmap.2
}
