% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfishMapping.r
\name{rotateXY}
\alias{rotateXY}
\title{Rotate coordinates}
\usage{
rotateXY(datFish, flatVector = NULL, flipVector = NULL,
  subset = NULL)
}
\arguments{
\item{datFish}{a fishScaleAndMap output list}

\item{flatVector}{a TRUE/FALSE vector ordred in the same way as the elements (e.g., cells)
in datIn where all TRUE values correspond to cells who should have the same Y coordinate
(e.g., be in the same layer).  Alternatively a numeric vector of cell indices to include}

\item{flipVector}{a numeric vector of values to ensure proper reflection on Y-axes (e.g.,
layer; default=NULL)}

\item{subset}{a boolean or numeric vector of the elements to retain}
}
\value{
a fishScaleAndMap output list with updated scaledX and scaleY coordinates
}
\description{
Rotates the scaledX and scaledY elements of a fishScaleAndMap output list so that the
  axis of interest (e.g., cortical layer) is paralled with the x cooridate plan.
  Rotation code is from  https://stackoverflow.com/questions/15463462/rotate-graph-by-angle
}
