% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfishMapping.r
\name{summarizeMatrix}
\alias{summarizeMatrix}
\title{Summarize matrix}
\usage{
summarizeMatrix(mat, group, scale = "none", scaleQuantile = 1,
  binarize = FALSE, binMin = 0.5, summaryFunction = median, ...)
}
\arguments{
\item{mat}{matrix where the columns (e.g., samples) are going to be grouped}

\item{group}{vector of length dim(mat)[2] corresponding to the groups}

\item{scale}{either 'none' (default),'row', or 'column'}

\item{scaleQuantile}{what quantile of value should be set as 1 (default=1)}

\item{binarize}{should the data be binarized? (default=FALSE)}

\item{binMin}{minimum ON value for the binarized matrix (ignored if binarize=FALSE)}

\item{summaryFunction}{function (or function name) to be used for summarization}

\item{...}{additional parameters for summaryFunction}
}
\value{
matrix of summarized values
}
\description{
Groups columns in a matrix by a specified group vector and summarizes using a specificed function.
  Optionally binarizes the matrix using a specified cutoff parameter.  This is a wrapper for tapply.
}
